requirements.txt
pillow==10.3.0
torch==2.2.0
torchvision==0.17.2
numpy==1.26.4
scikit-learn==1.4.2
timm==0.4.9
tqdm==4.66.4
tensorflow==2.16.1
tensorboardX==2.6.2.2


